refactor(cuda): mul_mm's per-kind rows move to one module per family
`mul_mm.rs` was the GEMM body, the geometry constants, the emitter,
the shape refusals AND the six per-format unpack functions with their
six Rust twins, at 1,126 lines. Adding a quant kind meant opening it,
which is the reason given in CLAUDE.md for why llama.cpp has 140
architectures and this repo has 23 proven.

The rows now live in `mul_mm_kinds/`, one module per format family:
`legacy` (Q8_0, Q4_0, Q5_0 -- 32-element blocks, `nl == 2`) and
`kquant` (Q4_K, Q5_K, Q6_K -- 256-element super-blocks, two of them
sharing the 6-bit scale/min unpack). `mul_mm.rs` keeps the parts that
do not vary per format and falls to 664 lines.

Split by family rather than one file per kind because the families are
what actually share code, and the shared piece has to live somewhere
that is not "the file with everything in it".

No behaviour change. Every name stays where callers already spell it:
`mul_mm::Q4_K` and `mul_mm::K_SCALE_MIN_SRC` are re-exports now, so
`ferrox-core`, the host-check example and every test are untouched.
`KINDS` itself stays in `mul_mm`, so there is exactly one list and a
row added without a `KINDS` entry is dead code clippy names.

`cargo test -p ferrox-cuda`: 28 passed. Host check re-run after the
move: zero mismatches over 40,932 positions.
