+++ crates/ferrox-cuda/src/lib.rs
@@ -22,6 +22,10 @@ pub mod coalesced_twin;
 /// on a GPU-less host still exercises the arithmetic the kernel encodes.
 /// Only the launch path ([`mul_mm_launch`]) is feature-gated.
 pub mod mul_mm;
+/// One module per quant-format family, holding the [`mul_mm::KINDS`]
+/// rows. Split out of `mul_mm.rs` so that adding a format is a new file
+/// rather than a new section of a file nobody wants to open.
+pub mod mul_mm_kinds;
 pub mod mul_mm_ref;
 
 #[cfg(feature = "cuda")]
